requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==9.2.0
environs[django]==9.3.2
django-phonenumber-field==7.0.0
phonenumbers==8.12.56
djangorestframework==3.14.0
requests==2.26.0
geopy==2.2.0
rollbar==0.16.3
psycopg2==2.9.5
